Extract commonalities from Phone, Email and Address into parent class Contact
